requirements.dev.txt
aiohttp==3.9.4
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
attrs==23.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.5.14
deepdiff==7.0.1
distro==1.9.0
emoji==2.11.1
executing==2.0.1
faiss-cpu==1.7.4
filetype==1.2.0
Flask==3.0.2
Flask-Cors==4.0.0
frozenlist==1.4.1
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
icecream==2.1.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.4.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.46
langchain-openai==0.1.4
langchain-text-splitters==0.0.1
langchainplus-sdk==0.0.20
langdetect==1.0.9
langsmith==0.1.51
lxml==5.2.1
MarkupSafe==2.1.5
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
nltk==3.8.1
numexpr==2.10.0
numpy==1.26.4
openai==1.24.0
openapi-schema-pydantic==1.2.4
ordered-set==4.1.0
orjson==3.10.1
packaging==23.2
pydantic==1.10.15
pydantic_core==2.18.1
Pygments==2.17.2
pypdf==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-iso639==2024.2.7
python-magic==0.4.27
PyYAML==6.0.1
rapidfuzz==3.8.1
regex==2024.4.16
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.29
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.6.0
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.11.0
unstructured==0.13.4
unstructured-client==0.22.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
